move to pkg/client
